requirements.txt
autobahn_rce==0.6.0.5
Twisted==24.3.0
